# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"USER_STATUS": {
				"enum": [
					"active",
					"inactive",
					"disabled",
					"pending"
				],
				"type": "string"
			},
			"AccountVerifyToken": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"token": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"deletedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgottenPasswordToken": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"token": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"expiredOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"deletedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TwoFactorSession": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"token": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"userId",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TwoFactorAuthentication": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"secret": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GeneratedSecret"
							},
							{
								"$ref": "#/components/schemas/Record_string.never_"
							}
						],
						"nullable": true
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"userId",
					"secret"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Document": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"$ref": "#/components/schemas/DocumentType"
					},
					"ownerId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"ownerType": {
						"$ref": "#/components/schemas/DocumentOwnerType"
					},
					"mimeType": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"properties": {
						"$ref": "#/components/schemas/DocumentProperties"
					},
					"companyId": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"approved": {
						"type": "boolean"
					},
					"stored": {
						"properties": {
							"thumbnail": {
								"type": "string"
							},
							"resized": {
								"type": "string"
							},
							"original": {
								"type": "string"
							}
						},
						"required": [
							"original"
						],
						"type": "object"
					}
				},
				"required": [
					"type",
					"mimeType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tenant": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"status": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TENANT_STATUS.ACTIVE"
							},
							{
								"$ref": "#/components/schemas/TENANT_STATUS.INACTIVE"
							}
						]
					},
					"taxNumber": {
						"type": "string"
					},
					"registrationNumber": {
						"type": "string"
					},
					"bankAccount": {
						"type": "string",
						"nullable": true
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/Document"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"country",
					"city",
					"address",
					"zipCode",
					"status",
					"taxNumber",
					"registrationNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					},
					"tenant": {
						"$ref": "#/components/schemas/Tenant"
					},
					"tenantId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Notifications": {
				"properties": {
					"userNew": {
						"type": "boolean"
					},
					"tenderNew": {
						"type": "boolean"
					},
					"contactNew": {
						"type": "boolean"
					},
					"userDelete": {
						"type": "boolean"
					},
					"userUpdate": {
						"type": "boolean"
					},
					"customerNew": {
						"type": "boolean"
					},
					"contactUpdate": {
						"type": "boolean"
					},
					"contractorNew": {
						"type": "boolean"
					},
					"customerUpdate": {
						"type": "boolean"
					},
					"tenderApproved": {
						"type": "boolean"
					},
					"userUpdateRole": {
						"type": "boolean"
					},
					"contractorUpdate": {
						"type": "boolean"
					},
					"tenderStatusChange": {
						"type": "boolean"
					},
					"permissionMatrixUpdate": {
						"type": "boolean"
					},
					"tenderAwaitingApproval": {
						"type": "boolean"
					}
				},
				"required": [
					"userNew",
					"tenderNew",
					"contactNew",
					"userDelete",
					"userUpdate",
					"customerNew",
					"contactUpdate",
					"contractorNew",
					"customerUpdate",
					"tenderApproved",
					"userUpdateRole",
					"contractorUpdate",
					"tenderStatusChange",
					"permissionMatrixUpdate",
					"tenderAwaitingApproval"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Salary": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"hourlyRate": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"dailyRate": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Contact": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"phoneNumber": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/CONTACT_STATUS.ACTIVE"
							},
							{
								"$ref": "#/components/schemas/CONTACT_STATUS.INACTIVE"
							}
						]
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"user": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"tenantId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"Partial_User-or-null_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"salt": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/USER_STATUS"
					},
					"enableTwoFactor": {
						"type": "boolean"
					},
					"country": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"deletedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					},
					"deletedBy": {
						"type": "number",
						"format": "double"
					},
					"accountVerifyToken": {
						"$ref": "#/components/schemas/AccountVerifyToken"
					},
					"accountVerifyTokenId": {
						"type": "number",
						"format": "double"
					},
					"forgottenPasswordToken": {
						"$ref": "#/components/schemas/ForgottenPasswordToken"
					},
					"forgottenPasswordTokenId": {
						"type": "number",
						"format": "double"
					},
					"twoFactorSession": {
						"$ref": "#/components/schemas/TwoFactorSession"
					},
					"twoFactorSessionId": {
						"type": "number",
						"format": "double"
					},
					"twoFactorAuthentication": {
						"$ref": "#/components/schemas/TwoFactorAuthentication"
					},
					"twoFactorAuthenticationId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"tenant": {
						"$ref": "#/components/schemas/Tenant"
					},
					"tenantId": {
						"type": "number",
						"format": "double"
					},
					"lastLoggedIn": {
						"type": "string",
						"format": "date-time"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/Document"
						},
						"type": "array"
					},
					"notifications": {
						"$ref": "#/components/schemas/Notifications"
					},
					"entity": {
						"type": "string"
					},
					"enableLogin": {
						"type": "boolean"
					},
					"identifier": {
						"type": "string"
					},
					"employeeType": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"birthName": {
						"type": "string"
					},
					"motherName": {
						"type": "string"
					},
					"placeOfBirth": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"socialSecurityNumber": {
						"type": "string"
					},
					"taxIdentificationNumber": {
						"type": "string"
					},
					"personalIdentificationNumber": {
						"type": "string"
					},
					"licensePlateNumber": {
						"type": "string"
					},
					"salaries": {
						"items": {
							"$ref": "#/components/schemas/Salary"
						},
						"type": "array"
					},
					"contact": {
						"$ref": "#/components/schemas/Contact"
					},
					"contactId": {
						"type": "number",
						"format": "double"
					},
					"properties": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"billing": {
						"anyOf": [
							{
								"properties": {},
								"type": "object"
							},
							{
								"$ref": "#/components/schemas/UserBilling"
							}
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"LoginResponse": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"twoFactorAuthenticationEnabled": {
						"type": "boolean"
					},
					"session": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string",
						"pattern": "^(.+)@(.+)$"
					},
					"password": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Login2FaRequest": {
				"properties": {
					"token": {
						"type": "string"
					},
					"session": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"session"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyAccountResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyAccountRequest": {
				"properties": {
					"password": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ForgottenPasswordResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgottenPasswordRequest": {
				"properties": {
					"email": {
						"type": "string",
						"pattern": "^(.+)@(.+)$"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"properties": {
					"password": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Vilkrig x Alpinnovacio CRM API",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {}
	},
	"paths": {
...
		"/auth/login": {
			"post": {
				"operationId": "LogIn",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"description": "Authenticates a user and initiates a session. This endpoint is responsible for\nvalidating user credentials (like email and password) and, if successful,\nissuing authentication tokens.\n\nOn successful authentication, it sets the necessary cookies for maintaining\nthe user's session. These include a token and a refresh token, which are\nused for maintaining session state and refreshing the session, respectively.\n\nThe actual authentication logic is managed by the authentication service,\nwhile this method serves as an interface between the HTTP request and the\nservice layer.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The login request body containing the user's login credentials.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest",
								"description": "The login request body containing the user's login credentials."
							}
						}
					}
				}
			}
		},
		"/auth/login/2fa": {
			"post": {
				"operationId": "LogIn2Fa",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"description": "Authenticates a user with two-factor authentication (2FA). This endpoint extends the basic login \nfunctionality by adding a second layer of security. After the initial credentials (like email and password) \nare validated, this method requires a second verification step, typically a code from a device or app.\n\nSimilar to the basic login, upon successful authentication, it sets up the necessary cookies. However,\nthe focus here is on securely managing the 2FA process, ensuring that both the primary credentials \nand the 2FA verification are successful before issuing tokens.\n\nThis method also serves as a bridge between the HTTP request and the authentication service, \nspecifically handling the 2FA logic. It forwards the context and the 2FA credentials (like a verification code)\nto the service layer for processing.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The 2FA request body, containing both the user's initial login credentials and the 2FA verification code.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login2FaRequest",
								"description": "The 2FA request body, containing both the user's initial login credentials and the 2FA verification code."
							}
						}
					}
				}
			}
		},
		"/auth/verify-account": {
			"post": {
				"operationId": "VerifyAccount",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyAccountResponse"
								}
							}
						}
					}
				},
				"description": "Processes user account verification requests. This endpoint is used to verify\na user's account typically after they have signed up. The verification is\nusually done using a token that is sent to the user's email.\n\nThe method takes a verification token as a query parameter and other relevant\ndata in the request body. It then interacts with the authentication service to\nvalidate the token and complete the verification process.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "The verification token, usually passed as a query parameter.",
						"in": "query",
						"name": "v",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The request body containing any additional data needed for verification.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyAccountRequest",
								"description": "The request body containing any additional data needed for verification."
							}
						}
					}
				}
			}
		},
...
		"/auth/forgotten-password": {
			"post": {
				"operationId": "ForgottenPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForgottenPasswordResponse"
								}
							}
						}
					}
				},
				"description": "Handles forgotten password requests by initiating a password reset process.\nThis method is responsible for sending a password reset link or code to the\nuser's registered email address. It is an essential part of user account\nmanagement, especially in scenarios where a user has forgotten their password\nand needs to reset it.\n\nThe method accepts the user's email (or other identifying information) as part\nof the request body and passes this information to the authentication service\nto handle the sending of the reset link or code.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The request body containing the necessary data for processing\nforgotten password requests, such as the user's email.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgottenPasswordRequest",
								"description": "The request body containing the necessary data for processing\nforgotten password requests, such as the user's email."
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordResponse"
								}
							}
						}
					}
				},
				"description": "Manages the password reset process for users. This endpoint is invoked when a\nuser wants to reset their password, typically following a forgotten password\nscenario. It requires a unique token, usually sent to the user's email, along\nwith their new password.\n\nThe token ensures that the request is legitimate and is often passed as a query\nparameter. The method verifies the token's validity and, if successful, updates\nthe user's password with the new one provided in the request body.\n\nThis method is a key component in ensuring user account security, allowing users\nto regain access to their account securely after losing their password.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "The verification token, typically provided as a query parameter,\nused to validate the password reset request.",
						"in": "query",
						"name": "v",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The request body, containing the new password and any additional\ndata required for the password reset.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest",
								"description": "The request body, containing the new password and any additional\ndata required for the password reset."
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_User-or-null_"
								}
							}
						}
					}
				},
				"description": "Retrieves the authenticated user's profile information. This endpoint is typically\ncalled after a user has logged in and received an authentication token. It's used\nto fetch the user's data such as name, email, and any other relevant profile information.\n\nThe method requires an authentication token, usually provided in the request headers,\nto verify the user's identity. Upon successful authentication, it returns the user's\nprofile data. This ensures that only authenticated users can access their own profile\ninformation, maintaining privacy and security.\n\nThis method is crucial for personalized user experiences, allowing the system to\nidentify and provide specific information relevant to the logged-in user.",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwtToken": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}